refactor: screenをpagesに移動
https://note.com/tabelog_frontend/n/n07b4077f5cf3#5WquU
