+++ src/components/pages/CheckinDetail.tsx
