+++ src/components/pages/Onboarding.tsx
@@ -3,6 +3,7 @@ import { StyleSheet } from 'react-native'
 import Onboarding from 'react-native-onboarding-swiper'
 import { Image } from 'react-native-elements'
 import { useNavigation } from '@react-navigation/native'
+import CalendarImage from '../../../assets/images/calendar-view.png'
 
 const AppOnboarding = () => {
   const navigation = useNavigation()
@@ -14,7 +15,7 @@ const AppOnboarding = () => {
       pages={[
         {
           backgroundColor: '#fff',
-          image: <Image source={require('../../assets/images/calendar-view.png')} />,
+          image: <Image source={CalendarImage} />,
           title: 'SwarmPlus',
           subtitle: 'あの日を振り返ろう',
         },
