+++ src/navigation/index.tsx
@@ -7,8 +7,8 @@ import { RootStackParamList } from '@/types'
 import BottomTabNavigator from '@/navigation/BottomTabNavigator'
 import LinkingConfiguration from '@/navigation/LinkingConfiguration'
 import { DarkTheme, DefaultTheme } from '@/constants/Theme'
-import AppOnboarding from '@/screens/Onboarding'
-import SignInByFoursquare from '@/screens/SignInByFoursquare'
+import AppOnboarding from '@/components/pages/Onboarding'
+import SignInByFoursquare from '@/components/pages/SignInByFoursquare'
 import { useState, useEffect, useRef } from 'react'
 import storage from '@/service/reactNativeStorage'
 import { FOURSQUARE_ACCESS_TOKEN } from '../constants/StorageKeys'
